backend/requirements.txt
# Shiksha Saathi - Backend Requirements
# Python 3.10+ | Conda env: prajatantra

# ═══════════════════════════════════════════════════════════════════════════════
# DJANGO & DRF
# ═══════════════════════════════════════════════════════════════════════════════
Django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# ═══════════════════════════════════════════════════════════════════════════════
# ENVIRONMENT & CONFIG
# ═══════════════════════════════════════════════════════════════════════════════
python-dotenv>=1.0.0

# ═══════════════════════════════════════════════════════════════════════════════
# GEMINI AI
# ═══════════════════════════════════════════════════════════════════════════════
google-generativeai>=0.8.3

# ═══════════════════════════════════════════════════════════════════════════════
# RAG PIPELINE (Migrated from Flask)
# ═══════════════════════════════════════════════════════════════════════════════
chromadb>=0.4.22
pypdf>=4.0.1
sentence-transformers>=2.3.1
numpy>=1.26.4

# ═══════════════════════════════════════════════════════════════════════════════
# YOUTUBE SEARCH
# ═══════════════════════════════════════════════════════════════════════════════
youtube-search-python>=1.6.6
httpx>=0.26.0

# ═══════════════════════════════════════════════════════════════════════════════
# UTILITIES
# ═══════════════════════════════════════════════════════════════════════════════
tenacity>=8.2.3
